+++ src/core/button/demo/index.tsx
@@ -1,39 +1,57 @@
 import React, { useState } from 'react';
-import { Avatar } from '..';
+import { Button } from '..';
 
 export default () => {
   const defaultAvatarSrc = "https://www.jiuwa.net/uploads/image/20170506/20170506084900_43824.jpg";
   const [avatarSrc, setAvatarSrc] = useState(defaultAvatarSrc);
   return (
     <div>
-      <Avatar
-        size={100}
-        onChangeAvatar={(e) => {
-          { /* alert('你好') */ }
-          console.log(e);
-        }}
-        tip={10}
-        changeAvatarable
-        text="A" />
+      <Button
+        className="mr10"
+        onClick={(e) => {
+          ShowModal({
+            title: '试试',
+            children: (
+              <div className="p20">你好</div>
+            )
+          });
+        }} icon="archway">
+        弹窗
+      </Button>
       <hr />
-      <Avatar
-        size={30}
-        text="A" />
-      <hr />
-      <Avatar
-        size={100}
-        onChangeAvatar={(res) => {
-          console.log(res);
-          setAvatarSrc(res);
+      <Button
+        onClick={(e) => {
+          setState({
+            loading: !state.loading
+          });
+          setTimeout(() => {
+            setState(({ loading }) => ({
+              loading: !loading
+            }));
+          }, 1500);
         }}
-        changeAvatarable
-        position="top"
-        faceOptions={[
-          'https://tu.jiuwa.net/bg/992.jpg',
-          'https://is2-ssl.mzstatic.com/image/thumb/Purple1/v4/b4/87/68/b487686b-27ec-9bb2-4fff-0c924c426a01/mzl.osfqhmvz.jpg/246x0w.jpg',
-          'https://www.jiuwa.net/uploads/image/20170506/20170506084900_43824.jpg',
-        ]}
-        src={avatarSrc} />
+        loading={state.loading}
+        icon="bolt"
+        color="green"
+        text="试试切换 Loading" />
+      <hr />
+      <Button
+        onClick={(e) => {
+          alert('你好');
+        }}
+        icon="bell-slash"
+        color="red"
+        disabled
+        text="禁用按钮" />
+      <hr />
+      <Button
+        onClick={(e) => {
+          alert('你好');
+        }}
+        icon="bell-slash"
+        className="res"
+        disabled
+        text="响应式按钮" />
     </div>
   );
 };