+++ src/core/utils/color.ts
@@ -1,15 +1,29 @@
+export const ExampleColors = [
+  // nature color
+  // 'theme',
+  'blue',
+  'green',
+  'orange',
+  'red',
+  'default',
+  // 'cyan',
+  // 'purple',
+  // 'wine',
+  // 'yellow',
+  // 'gold',
+];
 export const NatureColors = [
   // nature color
   'theme',
-  'green',
   'blue',
-  'red',
-  'yellow',
-  'cyan',
-  'gold',
-  'purple',
+  'green',
   'orange',
+  'red',
+  'cyan',
+  'purple',
   'wine',
+  'yellow',
+  'gold',
 ];
 export const GrayscaleColors = [
   // Grayscale
@@ -22,6 +36,7 @@ export const StatusColors = [
   // status color
   'important',
   'primary',
+  'danger',
   'error',
   'success',
   'warn',
@@ -29,12 +44,12 @@ export const StatusColors = [
 ];
 export const ColorGroup = [...NatureColors, ...GrayscaleColors, ...StatusColors];
 
-type ForEachColorParams = (color: string) => JSX.Element | JSX.Element[];
+type ForEachColorParams = (color: string, idx: number) => JSX.Element | JSX.Element[];
 /**
  * 遍历颜色系统选项
  * @param Colors
  */
-export const ForEachColor = (Colors = ColorGroup) => (getComponent: ForEachColorParams) => {
+export const ForEachColor = (Colors = ExampleColors) => (getComponent: ForEachColorParams) => {
   if (!getComponent) return null;
-  return Colors.map(color => getComponent(color));
+  return Colors.map((color, idx) => getComponent(color, idx));
 };